+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Changed
+* Changed imported material names to use the model name as a prefix to avoid name conflicts across imports.
+
 ## 0.17.0 - 2024-12-20
 ### Added
 * Added support for importing collisions from .wiidcm or .idcm files.