+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unreleased
+## 0.18.0 - 2025-01-14
 ### Added
 * Added an option to merge wimdo and camdo armatures on import. Leave this unchecked for editing and exporting models.